refactor: tobe.method -> tobe.resolver 패키지명 변경
