+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/resolver/DefaultHandlerMethodArgumentResolver.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe.method;
+package nextstep.mvc.tobe.resolver;
 
 
 import nextstep.mvc.tobe.exception.HandlerMethodArgumentResolverException;
