+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/resolver/PathVariableHandlerMethodArgumentResolver.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe.method;
+package nextstep.mvc.tobe.resolver;
 
 import nextstep.utils.PathUtils;
 import nextstep.web.annotation.PathVariable;
